palyer and queue as components
